+++ e2e/imgbuild/imgbuild.go
@@ -13,11 +13,9 @@ import (
 	"os/exec"
 	"path"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/sylabs/singularity/e2e/internal/e2e"
-	"github.com/sylabs/singularity/internal/pkg/util/bin"
 )
 
 var testFileContent = "Test file content\n"
@@ -25,10 +23,6 @@ var testFileContent = "Test file content\n"
 type imgBuildTests struct {
 	env e2e.TestEnv
 }
-
-const (
-	passphrase = "e2e-passphrase"
-)
 
 func (c *imgBuildTests) buildFrom(t *testing.T) {
 	e2e.PrepRegistry(t, c.env)
@@ -65,7 +59,7 @@ func (c *imgBuildTests) buildFrom(t *testing.T) {
 
 		c.env.RunSingularity(
 			t,
-			e2e.WithPrivileges(true),
+			e2e.WithProfile(e2e.RootProfile),
 			e2e.WithCommand("build"),
 			e2e.WithArgs(args...),
 			e2e.PreRun(func(t *testing.T) {
@@ -141,7 +135,7 @@ func (c *imgBuildTests) nonRootBuild(t *testing.T) {
 
 		c.env.RunSingularity(
 			t,
-			e2e.WithPrivileges(false),
+			e2e.WithProfile(e2e.UserProfile),
 			e2e.WithCommand("build"),
 			e2e.WithArgs(args...),
 			e2e.PostRun(func(t *testing.T) {
@@ -183,7 +177,7 @@ func (c *imgBuildTests) buildLocalImage(t *testing.T) {
 
 	c.env.RunSingularity(
 		t,
-		e2e.WithPrivileges(true),
+		e2e.WithProfile(e2e.RootProfile),
 		e2e.WithCommand("build"),
 		e2e.WithArgs("--sandbox", sandboxImage, c.env.ImagePath),
 		e2e.PostRun(func(t *testing.T) {
@@ -215,7 +209,7 @@ func (c *imgBuildTests) buildLocalImage(t *testing.T) {
 		c.env.RunSingularity(
 			t,
 			e2e.AsSubtest(tt.name),
-			e2e.WithPrivileges(true),
+			e2e.WithProfile(e2e.RootProfile),
 			e2e.WithCommand("build"),
 			e2e.WithArgs(imagePath, tt.buildSpec),
 			e2e.PostRun(func(t *testing.T) {
@@ -230,7 +224,7 @@ func (c *imgBuildTests) badPath(t *testing.T) {
 	imagePath := path.Join(c.env.TestDir, "container")
 	c.env.RunSingularity(
 		t,
-		e2e.WithPrivileges(true),
+		e2e.WithProfile(e2e.RootProfile),
 		e2e.WithCommand("build"),
 		e2e.WithArgs(imagePath, "/some/dumb/path"),
 		e2e.ExpectExit(255),
@@ -426,7 +420,7 @@ func (c *imgBuildTests) buildMultiStageDefinition(t *testing.T) {
 
 		c.env.RunSingularity(
 			t,
-			e2e.WithPrivileges(true),
+			e2e.WithProfile(e2e.RootProfile),
 			e2e.WithCommand("build"),
 			e2e.WithArgs(args...),
 			e2e.PostRun(func(t *testing.T) {
@@ -723,7 +717,7 @@ func (c *imgBuildTests) buildDefinition(t *testing.T) {
 
 		c.env.RunSingularity(
 			t,
-			e2e.WithPrivileges(true),
+			e2e.WithProfile(e2e.RootProfile),
 			e2e.WithCommand("build"),
 			e2e.WithArgs(args...),
 			e2e.PostRun(func(t *testing.T) {
@@ -742,6 +736,7 @@ func (c *imgBuildTests) ensureImageIsEncrypted(t *testing.T, imgPath string) {
 	cmdArgs := []string{"info", sifID, imgPath}
 	c.env.RunSingularity(
 		t,
+		e2e.WithProfile(e2e.UserProfile),
 		e2e.WithCommand("sif"),
 		e2e.WithArgs(cmdArgs...),
 		e2e.ExpectExit(
@@ -751,36 +746,22 @@ func (c *imgBuildTests) ensureImageIsEncrypted(t *testing.T, imgPath string) {
 	)
 }
 
-func checkCryptsetupVersion() error {
-	cryptsetup, err := bin.Cryptsetup()
-	if err != nil {
-		return err
-	}
-
-	cmd := exec.Command(cryptsetup, "--version")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("failed to run cryptsetup --version: %s", err)
-	}
-
-	if !strings.Contains(string(out), "cryptsetup 2.") {
-		return fmt.Errorf("incompatible cryptsetup version")
-	}
-
-	return nil
-}
-
-// buildEncryptPassphrase is exercising the build command for encrypted containers
-// while using a passphrase. Note that it covers both the normal case and when the
-// version of cryptsetup available is not compliant.
-func (c *imgBuildTests) buildEncryptPassphrase(t *testing.T) {
+func (c *imgBuildTests) buildEncryptPemFile(t *testing.T) {
 	// Expected results for a successful command execution
 	expectedExitCode := 0
 	expectedStderr := ""
 
+	// We create a temporary directory to store the image, making sure tests
+	// will not pollute each other
+	tempDir, cleanup := e2e.MakeTempDir(t, c.env.TestDir, "", "")
+	defer cleanup(t)
+
+	// Generate the PEM file
+	pemFile, _ := e2e.GeneratePemFiles(t, c.env.TestDir)
+
 	// If the version of cryptsetup is not compatible with Singularity encryption,
 	// the build commands are expected to fail
-	err := checkCryptsetupVersion()
+	err := e2e.CheckCryptsetupVersion()
 	if err != nil {
 		expectedExitCode = 255
 		// todo: fix the problen with catching stderr, until then we do not do a real check
@@ -789,17 +770,13 @@ func (c *imgBuildTests) buildEncryptPassphrase(t *testing.T) {
 	}
 
 	// First with the command line argument
-	passphraseInput := []e2e.SingularityConsoleOp{
-		e2e.ConsoleSendLine(passphrase),
-	}
-	imgPath1 := filepath.Join(c.env.TestDir, "encrypted_cmdline_option.sif")
-	cmdArgs := []string{"-e", "--passphrase", imgPath1, "library://alpine:latest"}
+	imgPath1 := filepath.Join(tempDir, "encrypted_cmdline_option.sif")
+	cmdArgs := []string{"--encrypt", "--pem-path", pemFile, imgPath1, "library://alpine:latest"}
 	c.env.RunSingularity(
 		t,
+		e2e.WithProfile(e2e.RootProfile),
 		e2e.WithCommand("build"),
-		e2e.WithPrivileges(true),
 		e2e.WithArgs(cmdArgs...),
-		e2e.ConsoleRun(passphraseInput...),
 		e2e.ExpectExit(
 			expectedExitCode,
 			e2e.ExpectError(e2e.ContainMatch, expectedStderr),
@@ -811,15 +788,15 @@ func (c *imgBuildTests) buildEncryptPassphrase(t *testing.T) {
 	}
 
 	// Second with the environment variable
-	passphraseEnvVar := fmt.Sprintf("%s=%s", "SINGULARITY_ENCRYPTION_PASSPHRASE", passphrase)
-	imgPath2 := filepath.Join(c.env.TestDir, "encrypted_env_var.sif")
-	cmdArgs = []string{"-e", imgPath2, "library://alpine:latest"}
+	pemEnvVar := fmt.Sprintf("%s=%s", "SINGULARITY_ENCRYPTION_PEM_PATH", pemFile)
+	imgPath2 := filepath.Join(tempDir, "encrypted_env_var.sif")
+	cmdArgs = []string{"--encrypt", imgPath2, "library://alpine:latest"}
 	c.env.RunSingularity(
 		t,
+		e2e.WithProfile(e2e.RootProfile),
 		e2e.WithCommand("build"),
 		e2e.WithArgs(cmdArgs...),
-		e2e.WithPrivileges(true),
-		e2e.WithEnv(append(os.Environ(), passphraseEnvVar)),
+		e2e.WithEnv(append(os.Environ(), pemEnvVar)),
 		e2e.ExpectExit(
 			expectedExitCode,
 			e2e.ExpectError(e2e.ContainMatch, expectedStderr),
@@ -831,8 +808,111 @@ func (c *imgBuildTests) buildEncryptPassphrase(t *testing.T) {
 	}
 }
 
-// RunE2ETests is the main func to trigger the test suite
-func RunE2ETests(env e2e.TestEnv) func(*testing.T) {
+// buildEncryptPassphrase is exercising the build command for encrypted containers
+// while using a passphrase. Note that it covers both the normal case and when the
+// version of cryptsetup available is not compliant.
+func (c *imgBuildTests) buildEncryptPassphrase(t *testing.T) {
+	// Expected results for a successful command execution
+	expectedExitCode := 0
+	expectedStderr := ""
+
+	// We create a temporary directory to store the image, making sure tests
+	// will not pollute each other
+	tempDir, cleanup := e2e.MakeTempDir(t, c.env.TestDir, "", "")
+	defer cleanup(t)
+
+	// If the version of cryptsetup is not compatible with Singularity encryption,
+	// the build commands are expected to fail
+	err := e2e.CheckCryptsetupVersion()
+	if err != nil {
+		expectedExitCode = 255
+		expectedStderr = ": available cryptsetup is not supported"
+	}
+
+	// First with the command line argument, only using --passphrase
+	passphraseInput := []e2e.SingularityConsoleOp{
+		e2e.ConsoleSendLine(e2e.Passphrase),
+	}
+	cmdlineTestImgPath := filepath.Join(tempDir, "encrypted_cmdline_option.sif")
+	// The image is deleted during cleanup of the tempdir
+	cmdArgs := []string{"--passphrase", cmdlineTestImgPath, "library://alpine:latest"}
+	c.env.RunSingularity(
+		t,
+		e2e.AsSubtest("passphrase flag"),
+		e2e.WithProfile(e2e.RootProfile),
+		e2e.WithCommand("build"),
+		e2e.WithArgs(cmdArgs...),
+		e2e.ConsoleRun(passphraseInput...),
+		e2e.ExpectExit(
+			expectedExitCode,
+			e2e.ExpectError(e2e.ContainMatch, expectedStderr),
+		),
+	)
+	// If the command was supposed to succeed, we check the image
+	if expectedExitCode == 0 {
+		c.ensureImageIsEncrypted(t, cmdlineTestImgPath)
+	}
+
+	// With the command line argument, using --encrypt and --passphrase
+	cmdlineTest2ImgPath := filepath.Join(tempDir, "encrypted_cmdline2_option.sif")
+	cmdArgs = []string{"--encrypt", "--passphrase", cmdlineTest2ImgPath, "library://alpine:latest"}
+	c.env.RunSingularity(
+		t,
+		e2e.AsSubtest("encrypt and passphrase flags"),
+		e2e.WithProfile(e2e.RootProfile),
+		e2e.WithCommand("build"),
+		e2e.WithArgs(cmdArgs...),
+		e2e.ConsoleRun(passphraseInput...),
+		e2e.ExpectExit(
+			expectedExitCode,
+			e2e.ExpectError(e2e.ContainMatch, expectedStderr),
+		),
+	)
+	// If the command was supposed to succeed, we check the image
+	if expectedExitCode == 0 {
+		c.ensureImageIsEncrypted(t, cmdlineTest2ImgPath)
+	}
+
+	// With the environment variable
+	passphraseEnvVar := fmt.Sprintf("%s=%s", "SINGULARITY_ENCRYPTION_PASSPHRASE", e2e.Passphrase)
+	envvarImgPath := filepath.Join(tempDir, "encrypted_env_var.sif")
+	cmdArgs = []string{"--encrypt", envvarImgPath, "library://alpine:latest"}
+	c.env.RunSingularity(
+		t,
+		e2e.AsSubtest("passphrase env var"),
+		e2e.WithProfile(e2e.RootProfile),
+		e2e.WithCommand("build"),
+		e2e.WithArgs(cmdArgs...),
+		e2e.WithEnv(append(os.Environ(), passphraseEnvVar)),
+		e2e.ExpectExit(
+			expectedExitCode,
+			e2e.ExpectError(e2e.ContainMatch, expectedStderr),
+		),
+	)
+	// If the command was supposed to succeed, we check the image
+	if expectedExitCode == 0 {
+		c.ensureImageIsEncrypted(t, envvarImgPath)
+	}
+
+	// Finally a test that must fail: try to specify the passphrase on the command line
+	dummyImgPath := filepath.Join(tempDir, "dummy_encrypted_env_var.sif")
+	cmdArgs = []string{"--encrypt", "--passphrase", e2e.Passphrase, dummyImgPath, "library://alpine:latest"}
+	c.env.RunSingularity(
+		t,
+		e2e.AsSubtest("passphrase on cmdline"),
+		e2e.WithProfile(e2e.RootProfile),
+		e2e.WithCommand("build"),
+		e2e.WithArgs(cmdArgs...),
+		e2e.WithEnv(append(os.Environ(), passphraseEnvVar)),
+		e2e.ExpectExit(
+			1,
+			e2e.ExpectError(e2e.RegexMatch, `^Error for command \"build\": accepts 2 arg\(s\), received 3`),
+		),
+	)
+}
+
+// E2ETests is the main func to trigger the test suite
+func E2ETests(env e2e.TestEnv) func(*testing.T) {
 	c := &imgBuildTests{
 		env: env,
 	}
@@ -852,5 +932,6 @@ func RunE2ETests(env e2e.TestEnv) func(*testing.T) {
 		t.Run("MultiStage", c.buildMultiStageDefinition)
 		// build encrypted images
 		t.Run("buildEncryptPassphrase", c.buildEncryptPassphrase)
+		t.Run("buildEncryptPemFile", c.buildEncryptPemFile)
 	}
 }