+++ internal/app/starter/rpc_linux.go
@@ -9,11 +9,11 @@ import (
 	"net"
 	"os"
 
-	"github.com/sylabs/singularity/internal/pkg/runtime/engines/engine"
+	"github.com/sylabs/singularity/internal/pkg/runtime/engine"
 	"github.com/sylabs/singularity/internal/pkg/sylog"
 )
 
-// RPCServer serves runtime engine requests
+// RPCServer serves runtime engine requests.
 func RPCServer(socket int, e *engine.Engine) {
 	comm := os.NewFile(uintptr(socket), "unix")
 	conn, err := net.FileConn(comm)